[REMANIEMENT] Sors le tri des indicateurs de la classe `AdaptateurDeRestitution`
… pour la placer plus proche du type portant les indicateurs.
Notre objectif est d'alléger la classe abstraite `AdaptateurDeRestitution` qui fait un niveau d'héritage, donc de complexité, dont on pourrait se passer.
