+++ mon-aide-cyber-api/src/adaptateurs/AdaptateurDeRestitution.ts
@@ -1,8 +1,4 @@
-import {
-  Indicateurs,
-  MesurePriorisee,
-  ORDRE_THEMATIQUES,
-} from '../diagnostic/Diagnostic';
+import { MesurePriorisee } from '../diagnostic/Diagnostic';
 import { Restitution } from '../restitution/Restitution';
 
 export const estMesurePrioritaire = (
@@ -21,24 +17,4 @@ export abstract class AdaptateurDeRestitution<T> {
   constructor(protected readonly traductionThematiques: Map<string, string>) {}
 
   public abstract genereRestitution(restitution: Restitution): Promise<T>;
-
-  protected trieLesIndicateurs(restitution: Restitution) {
-    const indicateursRestitution: Indicateurs = Object.entries(
-      restitution.indicateurs
-    )
-      .sort(([thematiqueA], [thematiqueB]) =>
-        ORDRE_THEMATIQUES.indexOf(thematiqueA) >
-        ORDRE_THEMATIQUES.indexOf(thematiqueB)
-          ? 1
-          : -1
-      )
-      .reduce(
-        (accumulateur, [thematique, indicateur]) => ({
-          ...accumulateur,
-          [thematique]: indicateur,
-        }),
-        {}
-      );
-    return indicateursRestitution;
-  }
 }
