+++ mon-aide-cyber-api/src/restitution/Restitution.ts
@@ -1,6 +1,10 @@
 import { EntrepotEcriture } from '../domaine/Entrepot';
 import crypto from 'crypto';
-import { Indicateurs, Mesures } from '../diagnostic/Diagnostic';
+import {
+  Indicateurs,
+  Mesures,
+  ORDRE_THEMATIQUES,
+} from '../diagnostic/Diagnostic';
 
 type InformationsRestitution = {
   dateCreation: Date;
@@ -14,6 +18,27 @@ type Restitution = {
   informations: InformationsRestitution;
   mesures: Mesures;
 };
+
+export const trieLesIndicateurs = (restitution: Restitution) => {
+  const indicateursRestitution: Indicateurs = Object.entries(
+    restitution.indicateurs
+  )
+    .sort(([thematiqueA], [thematiqueB]) =>
+      ORDRE_THEMATIQUES.indexOf(thematiqueA) >
+      ORDRE_THEMATIQUES.indexOf(thematiqueB)
+        ? 1
+        : -1
+    )
+    .reduce(
+      (accumulateur, [thematique, indicateur]) => ({
+        ...accumulateur,
+        [thematique]: indicateur,
+      }),
+      {}
+    );
+  return indicateursRestitution;
+};
+
 type EntrepotRestitution = EntrepotEcriture<Restitution>;
 
 export { EntrepotRestitution, InformationsRestitution, Restitution };
